feat: 교수진 소개 페이지 msw 적용
