+++ apps/community/src/app/api/mock/professors/route.ts
@@ -1,0 +1,5 @@
+import { professors } from './data';
+
+export function GET() {
+  return Response.json({ data: professors });
+}
