+++ apps/community/src/app/member/professor/page.tsx
@@ -1,18 +1,23 @@
 import type { Metadata } from 'next';
 
+import { ProfessorCard } from '~/components/member/professor/professor-card';
 import { PageHeader } from '~/components/page-header';
-import { ProfessorCard } from '~/components/professor-card';
+
+import { getProfessors } from './remote';
 
 import * as styles from '~/app/member/professor/page.css';
-
-import { PROFESSORS } from './mocks/professor';
 
 export const metadata: Metadata = {
   title: '교수진 소개 - 경기대학교 AI컴퓨터공학부',
   description: '경기대학교 소프트웨어경영대학 AI컴퓨터공학부 공식 홈페이지',
 };
 
-export default function ProfessorPage() {
+//** TODO: for mocking */
+export const dynamic = 'force-dynamic';
+
+export default async function ProfessorPage() {
+  const { data } = await getProfessors();
+
   return (
     <section>
       <PageHeader
@@ -21,7 +26,7 @@ export default function ProfessorPage() {
       />
 
       <div className={styles.professorListWrapper}>
-        {PROFESSORS.map((professor) => (
+        {data.map((professor) => (
           <ProfessorCard
             key={`professor-${professor.id}`}
             professor={professor}
